+++ src/onion_balance.rs
@@ -10,7 +10,8 @@ use k8s_openapi::{
         },
     },
     apiextensions_apiserver::pkg::apis::apiextensions::v1::CustomResourceDefinition,
-    apimachinery::pkg::apis::meta::v1::LabelSelector,
+    apimachinery::pkg::apis::meta::v1::{Condition, LabelSelector, Time},
+    chrono::Utc,
 };
 use kube::{
     core::ObjectMeta,
@@ -22,8 +23,9 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
     kubernetes::{
-        self, error_policy, Annotations, Api, ConfigYaml, DeploymentContainerResources, Labels,
-        Object, Resource as KubernetesResource, ResourceName, SelectorLabels, Subset, Torrc,
+        self, error_policy, Annotations, Api, ConditionsExt, ConfigYaml,
+        DeploymentContainerResources, Labels, Object, Resource as KubernetesResource, ResourceName,
+        SelectorLabels, Subset, Torrc,
     },
     metrics::Metrics,
     onion_key::OnionKey,
@@ -44,7 +46,7 @@ use crate::{
 /// multiple hosts while also increasing resiliency by eliminating single
 /// points of failure.
 #[allow(clippy::module_name_repetitions)]
-#[derive(CustomResource, JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(CustomResource, JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[kube(
     group = "tor.agabani.co.uk",
     kind = "OnionBalance",
@@ -72,7 +74,7 @@ pub struct OnionBalanceSpec {
 }
 
 #[allow(clippy::module_name_repetitions)]
-#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OnionBalanceSpecConfigMap {
     /// Name of the Config Map.
@@ -82,7 +84,7 @@ pub struct OnionBalanceSpecConfigMap {
 }
 
 #[allow(clippy::module_name_repetitions)]
-#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OnionBalanceSpecDeployment {
     /// Containers of the Deployment.
@@ -95,7 +97,7 @@ pub struct OnionBalanceSpecDeployment {
 }
 
 #[allow(clippy::module_name_repetitions)]
-#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OnionBalanceSpecDeploymentContainers {
     /// Onion Balance container.
@@ -106,7 +108,7 @@ pub struct OnionBalanceSpecDeploymentContainers {
 }
 
 #[allow(clippy::module_name_repetitions)]
-#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OnionBalanceSpecDeploymentContainersOnionBalance {
     /// Resources of the container.
@@ -114,7 +116,7 @@ pub struct OnionBalanceSpecDeploymentContainersOnionBalance {
 }
 
 #[allow(clippy::module_name_repetitions)]
-#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OnionBalanceSpecDeploymentContainersTor {
     /// Resources of the container.
@@ -122,7 +124,7 @@ pub struct OnionBalanceSpecDeploymentContainersTor {
 }
 
 #[allow(clippy::module_name_repetitions)]
-#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OnionBalanceSpecOnionKey {
     /// Name of the OnionKey.
@@ -130,7 +132,7 @@ pub struct OnionBalanceSpecOnionKey {
 }
 
 #[allow(clippy::module_name_repetitions)]
-#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OnionBalanceSpecOnionService {
     /// OnionKey reference of the OnionService.
@@ -138,7 +140,7 @@ pub struct OnionBalanceSpecOnionService {
 }
 
 #[allow(clippy::module_name_repetitions)]
-#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OnionBalanceSpecOnionServiceOnionKey {
     /// Hostname value of the OnionKey.
@@ -148,9 +150,21 @@ pub struct OnionBalanceSpecOnionServiceOnionKey {
 }
 
 #[allow(clippy::module_name_repetitions)]
-#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[derive(JsonSchema, Deserialize, Serialize, Debug, Clone, PartialEq)]
 #[serde(rename_all = "camelCase")]
 pub struct OnionBalanceStatus {
+    /// Represents the latest available observations of a deployment's current state.
+    ///
+    /// ### Initialized
+    ///
+    /// `Initialized`
+    ///
+    /// ### OnionKey
+    ///
+    /// `NotFound`, `HostnameNotFound`, `Ready`
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub conditions: Vec<Condition>,
+
     /// OnionKey hostname.
     ///
     /// The hostname is only populated once `state` is "running".
@@ -158,15 +172,6 @@ pub struct OnionBalanceStatus {
 
     /// Number of OnionServices.
     pub onion_services: i32,
-
-    /// Human readable description of state.
-    ///
-    /// Possible values:
-    ///
-    /// - OnionKey not found
-    /// - OnionKey hostname not found
-    /// - running
-    pub state: String,
 }
 
 impl OnionBalance {
@@ -218,6 +223,11 @@ impl OnionBalance {
     #[must_use]
     pub fn onion_key_name(&self) -> ResourceName {
         (&self.spec.onion_key.name).into()
+    }
+
+    #[must_use]
+    pub fn status_conditions(&self) -> Option<&Vec<Condition>> {
+        self.status.as_ref().map(|f| f.conditions.as_ref())
     }
 }
 
@@ -325,15 +335,46 @@ impl kubernetes::Context for Context {
 enum State {
     OnionKeyNotFound,
     OnionKeyHostnameNotFound,
-    Running(OnionKey),
+    Initialized(OnionKey),
 }
 
-impl std::fmt::Display for State {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            State::OnionKeyNotFound => write!(f, "OnionKey not found"),
-            State::OnionKeyHostnameNotFound => write!(f, "OnionKey hostname not found"),
-            State::Running(_) => write!(f, "running"),
+impl From<&State> for Vec<Condition> {
+    fn from(value: &State) -> Self {
+        match value {
+            State::OnionKeyNotFound => vec![Condition {
+                last_transition_time: Time(Utc::now()),
+                message: "The OnionKey was not found.".into(),
+                observed_generation: None,
+                reason: "NotFound".into(),
+                status: "False".into(),
+                type_: "OnionKey".into(),
+            }],
+            State::OnionKeyHostnameNotFound => vec![Condition {
+                last_transition_time: Time(Utc::now()),
+                message: "The OnionKey does not have a hostname.".into(),
+                observed_generation: None,
+                reason: "HostnameNotFound".into(),
+                status: "False".into(),
+                type_: "OnionKey".into(),
+            }],
+            State::Initialized(_) => vec![
+                Condition {
+                    last_transition_time: Time(Utc::now()),
+                    message: "The OnionKey is ready.".into(),
+                    observed_generation: None,
+                    reason: "Ready".into(),
+                    status: "True".into(),
+                    type_: "OnionKey".into(),
+                },
+                Condition {
+                    last_transition_time: Time(Utc::now()),
+                    message: "The OnionBalance is initialized.".into(),
+                    observed_generation: None,
+                    reason: "Initialized".into(),
+                    status: "True".into(),
+                    type_: "Initialized".into(),
+                },
+            ],
         }
     }
 }
@@ -366,8 +407,8 @@ async fn reconciler(object: Arc<OnionBalance>, ctx: Arc<Context>) -> Result<Acti
     )
     .await?;
 
-    if let State::Running(onion_key) = &state {
-        // config map
+    if let State::Initialized(onion_key) = &state {
+        // ConfigMap
         reconcile_config_map(
             &Api::new(kube::Api::namespaced(ctx.client.clone(), &namespace)),
             &object,
@@ -378,7 +419,7 @@ async fn reconciler(object: Arc<OnionBalance>, ctx: Arc<Context>) -> Result<Acti
         )
         .await?;
 
-        // deployment
+        // Deployment
         reconcile_deployment(
             &Api::new(kube::Api::namespaced(ctx.client.clone(), &namespace)),
             &ctx.config,
@@ -402,7 +443,7 @@ async fn reconciler(object: Arc<OnionBalance>, ctx: Arc<Context>) -> Result<Acti
     tracing::info!("reconciled");
 
     match state {
-        State::Running(_) => Ok(Action::requeue(Duration::from_secs(3600))),
+        State::Initialized(_) => Ok(Action::requeue(Duration::from_secs(3600))),
         _ => Ok(Action::requeue(Duration::from_secs(5))),
     }
 }
@@ -418,7 +459,7 @@ async fn reconcile_onion_key(api: &Api<OnionKey>, object: &OnionBalance) -> Resu
         return Ok(State::OnionKeyHostnameNotFound);
     }
 
-    Ok(State::Running(onion_key))
+    Ok(State::Initialized(onion_key))
 }
 
 async fn reconcile_config_map(
@@ -477,13 +518,16 @@ async fn reconcile_onion_balance(
     api.update_status(
         object,
         OnionBalanceStatus {
-            hostname: if let State::Running(onion_key) = state {
+            conditions: object
+                .status_conditions()
+                .unwrap_or(&Vec::new())
+                .merge_from(&state.into()),
+            hostname: if let State::Initialized(onion_key) = state {
                 onion_key.hostname().map(Into::into)
             } else {
                 None
             },
             onion_services: i32::try_from(object.spec.onion_services.len()).unwrap(),
-            state: state.to_string(),
         },
     )
     .await